Clean up and secure open harness
