+++ src/systems/synthesis/service/src/lib/definitions/index.ts
@@ -4,4 +4,4 @@ export * from './implementation';
 export * from './model';
 export * from './models';
 export * from './provider';
-export * from './sandbox';
+export * from '../sandbox';
