+++ src/systems/synthesis/service/src/lib/open-harness/providers/node.ts
@@ -15,6 +15,8 @@ export interface NodeFsProviderOptions {
 }
 
 export class NodeFsProvider implements FsProvider {
+  readonly scope = 'host' as const;
+
   private cwd: string;
   private maxFileSize: number;
